TNTP pallette

This list of colors has been superseded by the new brand colors and the new function tntp_colors().

Usage

colors_tntp_palette

Format

An object of class character of length 16.

Examples

tntp_colors()
#>                mint|light_green gray|grey|light_gray|light_grey 
#>                       "#E2EDDC"                       "#F1F1EE" 
#>                           white                           black 
#>                       "#FFFFFF"                       "#000000" 
#>            yellow|medium_yellow                       navy|blue 
#>                       "#FDE57B"                       "#00355F" 
#>                  pink|light_red                             red 
#>                       "#FDDACA"                       "#C31F46" 
#>                           green    charcoal|dark_gray|dark_grey 
#>                       "#317D5C"                       "#4A4A4A" 
#>                tangerine|orange               salmon|medium_red 
#>                       "#F26C4C"                       "#DA8988" 
#>                            gold               moss|medium_green 
#>                       "#F2CF13"                       "#8FB09D" 
#>                  sky|light_blue            cerulean|medium_blue 
#>                       "#81D3EB"                       "#00A5C7" 
#>                    light_yellow         medium_gray|medium_grey 
#>                       "#FAEDB8"                       "#9E9E9C" 
#>                       dark_gold                      dark_green 
#>                       "#C2A60A"                       "#1D4935" 
#>                        dark_red                         green_4 
#>                       "#7C132C"                       "#60977D" 
#>                         green_2                          blue_4 
#>                       "#B9CFBD"                       "#006D93" 
#>                          blue_2                           red_4 
#>                       "#41BCD9"                       "#CF5467" 
#>                           red_2                        yellow_4 
#>                       "#ECB2A9"                       "#F8DA47" 
#>                        yellow_2                   gray_4|grey_4 
#>                       "#FCE99A"                       "#747473" 
#>                   gray_2|grey_2 
#>                       "#C7C7C5"