Convert all character vectors containing a set of values in a data.frame to factors.

This function examines each column in a data.frame; when it finds a column composed solely of the values provided to the lvls argument it updates them to be factor variables, with levels in the order provided.

This is an alternative to calling dplyr::mutate_at with factor() and identifying the specific variables you want to transform, if you have several repeated sets of responses.

Usage

factorize_df(dat, lvls)

Arguments

dat

data.frame with some factor variables stored as characters.

lvls

The factor levels in your variable(s), in order. If you have a question whose possible responses are a subset of another question's, don't use this function; manipulate the specific columns with dplyr::mutate_at.

Value

data.frame with factorization completed in place.

Examples

library(dplyr)
mtcars %>%
  dplyr::mutate(agr = rep(c("Somewhat agree", "Strongly disagree"), 16)) %>%
  factorize_df(lvls = c("Strongly disagree", "Somewhat disagree",
                        "Somewhat agree", "Strongly agree"))
#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if
#> `.name_repair` is omitted as of tibble 2.0.0.
#> ℹ Using compatibility `.name_repair`.
#> ℹ The deprecated feature was likely used in the tntpr package.
#>   Please report the issue to the authors.
#> Transformed these columns: 
#>  *  agr
#> # A tibble: 32 × 12
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb agr        
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <fct>      
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4 Somewhat a…
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4 Strongly d…
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1 Somewhat a…
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1 Strongly d…
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2 Somewhat a…
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1 Strongly d…
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4 Somewhat a…
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2 Strongly d…
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2 Somewhat a…
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4 Strongly d…
#> # ℹ 22 more rows

# prints warning due to case mismatches:
mtcars %>%
  dplyr::mutate(agr = rep(c("Somewhat Agree", "Strongly Disagree"), 16)) %>%
  factorize_df(lvls = c("Strongly disagree", "Somewhat disagree",
                        "Somewhat agree", "Strongly agree"))
#> Warning: No columns matched.  Check spelling & capitalization of your levels.
#> # A tibble: 32 × 12
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb agr        
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr>      
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4 Somewhat A…
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4 Strongly D…
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1 Somewhat A…
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1 Strongly D…
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2 Somewhat A…
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1 Strongly D…
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4 Somewhat A…
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2 Strongly D…
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2 Somewhat A…
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4 Strongly D…
#> # ℹ 22 more rows