Fake teacher roster dataset from OpenSDP

The Fake County synthetic panel dataset contains approximately 40,000 records comprising four years of data with roughly 10,000 teachers per year. The dataset includes information about teacher demographics, teaching assignments, salary, credentials, experience, evaluation scores, and hiring and retention status. It also includes information about school types and average student characteristics for each school. There are no real teachers in the dataset, but it is based on real data. Fake County was developed as an offshoot of the Strategic Data Project's work on human capital diagnostics for school districts and state education departments, and can be used for teaching or collaboration. The data was synthesized using the R synthpop package.

Usage

fake_county

Format

A data frame with 39,339 rows and 38 variables:

tid

double: Teacher ID

fake_data

double: Record Is Simulated

school_year

double: School Year

school_code

double: School Code

school_name

character: School Name

t_male

double: Teacher Is Male

t_race_ethnicity

double: Teacher Race/Ethnicity

t_job_area

double: Teacher Assignment Type

t_salary

double: Monthly Salary

t_nbpts

double: Teacher Has National Board Certification

t_tenured

double: Teacher Is Tenured

t_experience

double: Years of Teaching Experience

t_fte

double: Teacher's FTE Status

t_highest_degree

double: Teacher's Highest Degree

t_licensed_stem

double: Teacher Is Licensed In STEM Field

t_eval_obs

double: Evaluation Summary Observation Score

t_eval_growth

double: Evaluation Summary Student Growth Score

t_stay

double: Teacher in Same School in Following Year

t_transfer

double: Teacher in Different School in Following Year

t_leave

double: Teacher Not Teaching in Fake County Schools in Following Year

t_novice

double: Teacher Is Novice First-Year Teacher

t_new_hire

double: Teacher Did Not Teach in Fake County in Prior Year

sch_elem

double: School Is Elementary School

sch_middle

double: School Is Middle School

sch_high

double: School Is High School

sch_alternative

double: School Is Alternative School

sch_regular

double: School Is Regular School

sch_title_1

double: School Is Title 1 School

sch_magnet

double: School Is Magnet School

sch_vocational

double: School is Vocational School

sch_region

double: School Region Code

sch_calendar_type

double: School Calendar Type

sch_iep_pct

double: School Special Education Student Share in 2012-15

sch_minority_pct

double: School Minority Student Share in 2012-15

sch_frpl_pct

double: School Free and Reduced Price Lunch Student Share in 2012-15

sch_ela_avg

double: School ELA Test Score Average in 2012-15 (in standard deviations)

sch_math_avg

double: School Math Test Score Average in 2012-15 (in standard deviations)

sch_enroll_2015

double: School Enrollment in 2015

Source

https://github.com/OpenSDP/fake-county, posted under a Creative Commons license.