Fake teacher roster dataset from OpenSDP

The Fake County synthetic panel dataset contains approximately 40,000 records comprising four years of data with roughly 10,000 teachers per year. The dataset includes information about teacher demographics, teaching assignments, salary, credentials, experience, evaluation scores, and hiring and retention status. It also includes information about school types and average student characteristics for each school. There are no real teachers in the dataset, but it is based on real data. Fake County was developed as an offshoot of the Strategic Data Project's work on human capital diagnostics for school districts and state education departments, and can be used for teaching or collaboration. The data was synthesized using the R synthpop package.

Usage

fake_county

Format

A data frame with 39,339 rows and 38 variables:

\itemtiddouble: Teacher ID \itemfake_datadouble: Record Is Simulated \itemschool_yeardouble: School Year \itemschool_codedouble: School Code \itemschool_namecharacter: School Name \itemt_maledouble: Teacher Is Male \itemt_race_ethnicitydouble: Teacher Race/Ethnicity \itemt_job_areadouble: Teacher Assignment Type \itemt_salarydouble: Monthly Salary \itemt_nbptsdouble: Teacher Has National Board Certification \itemt_tenureddouble: Teacher Is Tenured \itemt_experiencedouble: Years of Teaching Experience \itemt_ftedouble: Teacher's FTE Status \itemt_highest_degreedouble: Teacher's Highest Degree \itemt_licensed_stemdouble: Teacher Is Licensed In STEM Field \itemt_eval_obsdouble: Evaluation Summary Observation Score \itemt_eval_growthdouble: Evaluation Summary Student Growth Score \itemt_staydouble: Teacher in Same School in Following Year \itemt_transferdouble: Teacher in Different School in Following Year \itemt_leavedouble: Teacher Not Teaching in Fake County Schools in Following Year \itemt_novicedouble: Teacher Is Novice First-Year Teacher \itemt_new_hiredouble: Teacher Did Not Teach in Fake County in Prior Year \itemsch_elemdouble: School Is Elementary School \itemsch_middledouble: School Is Middle School \itemsch_highdouble: School Is High School \itemsch_alternativedouble: School Is Alternative School \itemsch_regulardouble: School Is Regular School \itemsch_title_1double: School Is Title 1 School \itemsch_magnetdouble: School Is Magnet School \itemsch_vocati~ldouble: School is Vocational School \itemsch_regiondouble: School Region Code \itemsch_calendar_typedouble: School Calendar Type \itemsch_iep_pctdouble: School Special Education Student Share in 2012-15 \itemsch_minority_pctdouble: School Minority Student Share in 2012-15 \itemsch_frpl_pctdouble: School Free and Reduced Price Lunch Student Share in 2012-15 \itemsch_ela_avgdouble: School ELA Test Score Average in 2012-15 (in standard deviations) \itemsch_math_avgdouble: School Math Test Score Average in 2012-15 (in standard deviations) \itemsch_enrollment_2015double: School Enrollment in 2015

Source

https://github.com/OpenSDP/fake-county, posted under a Creative Commons license.